Animate the FAB before finishing the activity
Because of the shared element transition introduced earlier, now the FAB is
flickering when leaving the details screen. To fix this, the activity finish
is delayed by some about of millis to let the FAB complete the animation.
Now, FAB provides already a `hide()` method that should animate the view,
for some unknown reason however this method was not animating the FAB, that is
why a bit of custom animation was introduced.